requirements.txt
Flask
Flask-RESTful
Flask-Bcrypt
Flask-JWT-Extended
python-dotenv
pymongo
Werkzeug